update waf interface. update non working for now
